+++ lib/bali/path_helper.rb
@@ -7,10 +7,11 @@ module Bali
     # - exact: Matches the whole path
     # - partial: Matches if the path is included in the current path
     # - starts_with: Matches if the current path starts with the path
+    # - crud: Matches the HREF for any CRUD action.
     #
     # @param path [String] represents the path for the link
     # @param current_path [String] currently active URL
-    # @param match [Symbol] one of 3 match types [:exact, :partial, :starts_with]
+    # @param match [Symbol] one of 4 match types [:exact, :partial, :starts_with, :crud]
     def active_path?(path, current_path, match: :exact)
       return false if current_path.nil?
 
@@ -18,6 +19,9 @@ module Bali
       current_request_path = current_path.gsub(/\.html$/, '')
 
       case match
+      when :crud
+        path_without_params == current_request_path ||
+          show_or_edit_path?(path, current_request_path)
       when :starts_with
         current_request_path.starts_with?(path)
       when :partial
@@ -25,6 +29,12 @@ module Bali
       else
         path_without_params == current_request_path
       end
+    end
+
+    private
+
+    def show_or_edit_path?(path, current_request_path)
+      %r{\A#{path}/[[:digit:]]+(/edit)?\Z}.match(current_request_path)
     end
   end
 end